Refactor parameter names for clarity and consistency across various classes
- Updated parameter names in functions to improve readability and maintainability.
- Changed `shader_ptr` to `shader_ptr_in` in constructors of `Shape`, `QuadrilateralShape`, and `RectanglesShape`.
- Renamed `calendar_config_storage` to `incoming_calendar_config_storage` in `calendar_panel.hpp`.
- Updated `date_interval_bundles` to `incoming_date_interval_bundles` in `date_store.hpp` and related classes.
- Changed `title_config` to `incoming_title_config` in `title_panel.hpp` and `title_config.hpp`.
- Refactored `scene_graph.hpp`, `font.hpp`, `mvp_matrices.hpp`, and other files to use more descriptive parameter names.
- Ensured consistent use of `static_cast` for type conversions in various locations to enhance type safety.

+++ src/date_utils.hpp
@@ -143,7 +143,9 @@ inline boost::gregorian::date string_to_boost_date(const std::string &date_strin
     date_variable = boost::gregorian::date(boost::date_time::not_a_date_time);
   } else {
     try {
-      date_variable = boost::gregorian::date(date_parts[0], date_parts[1], date_parts[2]);
+      date_variable = boost::gregorian::date(static_cast<unsigned short>(date_parts[0]),
+                                             static_cast<unsigned short>(date_parts[1]),
+                                             static_cast<unsigned short>(date_parts[2]));
     } catch (const std::exception &) {
       date_variable = boost::gregorian::date(boost::date_time::not_a_date_time);
     }
